requirements.txt
streamlit==1.10.0
pandas==1.1.4
gspread==5.7.2
oauth2client==4.1.3
pytz==2022.7.1
langchain==0.0.94
dropbox==11.36.0
python-dotenv==1.0.0
